Clean resolvers
